requirements.txt
Django~=5.0.6
djangorestframework
boto3
feedparser
